pipeline configuration phase 1

+++ datawandcli/components/objects.py
@@ -79,6 +79,7 @@ class Base():
 class Configurable(Base):
     def __init__(self, name="", type="", path="", extensions=[]):
         super(Configurable, self).__init__(name, type, path, extensions)
+        self.is_clone = False
         self.config = {}
         
     @property
@@ -90,6 +91,18 @@ class Configurable(Base):
         if not isinstance(value, dict):
             raise ValueError("Parameters must be stored in a dictionary!")
         self._config = value
+        
+    def get(self, deps=[]):
+        conf = super(Configurable, self).get(deps)
+        conf["is_clone"] = "yes" if self.is_clone else "no"
+        if len(self.config) > 0:
+            conf["config"] = self.config
+        return conf
+        
+    def load(self, config):
+        super(Configurable, self).load(config)
+        self.is_clone = config["is_clone"] == "yes"
+        self.config = config.get("config",{})
         
 class ModuleObject(Configurable):
     def __init__(self, name="", type="", path="", extensions=["py"]):
@@ -108,6 +121,9 @@ class NotebookObject(Configurable):
         print("Creating new ipython notebook: %s" % path)
         create_notebook(path)
         
+    def copy(self):
+        return NotebookObject(self.name, self.type, self.path)
+        
 class PyScriptObject(Configurable):
     def __init__(self, name="", type="", path="", extensions=["py"]):
         super(PyScriptObject, self).__init__(name, type, path, extensions)
@@ -116,17 +132,19 @@ class PyScriptObject(Configurable):
         print("Creating new python script: %s" % path)
         create_pyscript(path)
         
+    def copy(self):
+        return PyScriptObject(self.name, self.type, self.path)
+        
 class Pipeline():
     def __init__(self, name="", directory=None, description=""):
         self.name = name
         self.description = description
         self.path = "%s/%s.json" % (directory, self.name)
-        self.modules = []
-        self.notebooks = []
-        self.pyscripts = []
-        self.dependencies = {}
         self.parts = {}
         self.file_paths = []
+        self.dependencies = {}
+        self.num_clones = {}
+        self.default_config = {}
 
     @property
     def name(self):
@@ -137,22 +155,30 @@ class Pipeline():
         if " " in str(value):
             raise ValueError("Pipeline name cannot contain spaces!")
         self._name = value
+        
+    @property
+    def modules(self):
+        return [obj for obj in self.parts.values() if isinstance(obj, ModuleObject)]
+    
+    @property
+    def notebooks(self):
+        return [obj for obj in self.parts.values() if isinstance(obj, NotebookObject)]
+    
+    @property
+    def pyscripts(self):
+        return [obj for obj in self.parts.values() if isinstance(obj, PyScriptObject)]
     
     @property
     def config(self):
         conf = {}
         conf["name"] = self.name
         conf["description"] = self.description
+        conf["default_config"] = self.default_config
         conf["config_path"] = self.path
         conf["imports"] = [item.get() for item in self.modules]
         conf["notebooks"] = [item.get(self.dependencies.get(item.name, [])) for item in self.notebooks]
         conf["py_scripts"] = [item.get(self.dependencies.get(item.name, [])) for item in self.pyscripts]
         return conf
-    
-    def clear(self, with_source=False):
-        names = list(self.parts.keys())
-        for obj_name in names:
-            self.remove(obj_name, with_source)
     
     def save(self):
         with open(self.path, 'w') as f:
@@ -166,19 +192,18 @@ class Pipeline():
                 config = json.load(f)
             self.name = config["name"]
             self.description = config["description"]
+            self.default_config = config["default_config"]
             self.path = config_path
+            self.num_clones = {}
             # parse modules
-            self.modules = []
             for item in config["imports"]:
                 mod = ModuleObject(item["name"], item["type"], item["path"])
                 self.add(mod)
             # parse notebooks
-            self.notebooks = []
             for item in config["notebooks"]:
                 nb = NotebookObject(item["name"], item["type"], item["path"])
                 self.add(nb)
             # parse scripts
-            self.pyscripts = []
             for item in config["py_scripts"]:
                 ps = PyScriptObject(item["name"], item["type"], item["path"])
                 self.add(ps)
@@ -187,6 +212,11 @@ class Pipeline():
                 deps = item.get("dependencies",[])
                 if len(deps) > 0:
                     self.add_dependencies(item["name"], deps)
+                if "CLONE" in item["name"]:
+                    name = item["name"].split("_CLONE")[0]
+                    if not name in self.num_clones:
+                        self.num_clones[name] = 0
+                    self.num_clones[name] += 1
             print("Pipeline was LOADED")
         else:
             raise ValueError("Invalid path! You must specify a JSON file.")
@@ -197,34 +227,21 @@ class Pipeline():
         if obj_name in self.parts:
             if not silent:
                 print("An object with 'name=%s' is already present in the pipeline. The names of the components must be unique!" % obj_name)
-        elif obj_path in self.file_paths:
-            if not silent:
-                print("An object with 'path=%s' is already present in the pipeline. Cannot add the same file twice!" % obj_path)
+        #elif obj_path in self.file_paths:
+        #    if not silent:
+        #        print("An object with 'path=%s' is already present in the pipeline. Cannot add the same file twice!" % obj_path)
         else:
-            if isinstance(obj, ModuleObject):
-                self.modules.append(obj)
-                self.parts[obj_name] = "module"
-            elif isinstance(obj, NotebookObject):
-                self.notebooks.append(obj)
-                self.parts[obj_name] = "notebook"
-            elif isinstance(obj, PyScriptObject):
-                self.pyscripts.append(obj)
-                self.parts[obj_name] = "pyscript"
-            else:
-                raise ValueError("Object with invalid type!")
+            self.parts[obj_name] = obj
             self.file_paths.append(obj_path)
     
     def remove(self, obj_name, with_source=False, silent=False):
         if obj_name in self.parts:
-            obj_type = self.parts[obj_name]
-            if obj_type == "module":
-                self.modules, fp = remove_from_pipeline(self.modules, obj_name)
-            elif obj_type == "notebook":
-                self.notebooks, fp = remove_from_pipeline(self.notebooks, obj_name)
-            else:
-                self.pyscripts, fp = remove_from_pipeline(self.pyscripts, obj_name)
+            obj = self.parts[obj_name]
+            fp = obj.path
             del self.parts[obj_name]
-            self.file_paths.remove(fp)
+            to_be_deleted = isinstance(obj, ModuleObject) or not obj.is_clone
+            if to_be_deleted:
+                self.file_paths.remove(fp)
             # remove all dependencies of this item
             if obj_name in self.dependencies:
                 del self.dependencies[obj_name]
@@ -233,7 +250,7 @@ class Pipeline():
             for key in keys:
                 if obj_name in self.dependencies[key]:
                     self.remove_dependencies(key, [obj_name])
-            if with_source:
+            if to_be_deleted and with_source:
                 os.remove(fp)
                 print("%s file was deleted!" % fp)
         else:
@@ -259,3 +276,26 @@ class Pipeline():
                         if len(self.dependencies[obj_name]) == 0:
                             del self.dependencies[obj_name]
                             break
+                            
+    def add_clone(self, obj_name, custom_config):
+        if obj_name in self.parts:
+            obj_config = self.default_config.copy()
+            obj_config.update(custom_config)
+            cnt = self.num_clones.get(obj_name, 0)
+            clone = self.parts[obj_name].copy()
+            clone.name = "%s_CLONE%i" % (obj_name, cnt+1)
+            clone.is_clone = True
+            clone.config = obj_config
+            self.add(clone)
+            print(obj_name, cnt+1)
+            self.num_clones[obj_name] = (cnt+1)
+        else:
+            raise ValueError("Invalid object name!")
+    
+    def clear(self):
+        items = list(self.num_clones.items())
+        for obj_name, cnt in items:
+            for i in range(cnt):
+                clone_name = "%s_CLONE%i" % (obj_name, i+1)
+                self.remove(clone_name, with_source=False)
+            del self.num_clones[obj_name]
